+++ color/src/tag.rs
@@ -4,8 +4,8 @@
 //! The color space tag enum.
 
 use crate::{
-    A98Rgb, AcesCg, ColorSpace, ColorSpaceLayout, DisplayP3, Hsl, Hwb, Lab, Lch, LinearSrgb,
-    Missing, Oklab, Oklch, ProphotoRgb, Rec2020, Srgb, XyzD50, XyzD65,
+    A98Rgb, Aces2065_1, AcesCg, ColorSpace, ColorSpaceLayout, DisplayP3, Hsl, Hwb, Lab, Lch,
+    LinearSrgb, Missing, Oklab, Oklch, ProphotoRgb, Rec2020, Srgb, XyzD50, XyzD65,
 };
 
 /// The color space tag for [dynamic colors].
@@ -13,6 +13,8 @@ use crate::{
 /// This represents a fixed set of known color spaces. The set is
 /// based on the CSS Color 4 spec, but might also extend to a small
 /// set of color spaces used in 3D graphics.
+///
+/// The integer values of these variants can change in breaking releases.
 ///
 /// [dynamic colors]: crate::DynamicColor
 //
@@ -46,13 +48,17 @@ pub enum ColorSpaceTag {
     ProphotoRgb = 10,
     /// The [`Rec2020`] color space.
     Rec2020 = 11,
+    /// The [`Aces2065_1`] color space.
+    Aces2065_1 = 15,
     /// The [`AcesCg`] color space.
     AcesCg = 12,
     /// The [`XyzD50`] color space.
     XyzD50 = 13,
     /// The [`XyzD65`] color space.
     XyzD65 = 14,
-    // NOTICE: If a new value is added, be sure to modify `MAX_VALUE` in the bytemuck impl.
+    // NOTICE: If a new value is added, be sure to modify `MAX_VALUE` in the bytemuck impl. Also
+    // note the variants' integer values are not necessarily in order, allowing newly added color
+    // space tags to be grouped with related color spaces.
 }
 
 impl ColorSpaceTag {
@@ -164,6 +170,7 @@ impl ColorSpaceTag {
             Self::A98Rgb => A98Rgb::from_linear_srgb(rgb),
             Self::ProphotoRgb => ProphotoRgb::from_linear_srgb(rgb),
             Self::Rec2020 => Rec2020::from_linear_srgb(rgb),
+            Self::Aces2065_1 => Aces2065_1::from_linear_srgb(rgb),
             Self::AcesCg => AcesCg::from_linear_srgb(rgb),
             Self::XyzD50 => XyzD50::from_linear_srgb(rgb),
             Self::XyzD65 => XyzD65::from_linear_srgb(rgb),
@@ -187,6 +194,7 @@ impl ColorSpaceTag {
             Self::A98Rgb => A98Rgb::to_linear_srgb(src),
             Self::ProphotoRgb => ProphotoRgb::to_linear_srgb(src),
             Self::Rec2020 => Rec2020::to_linear_srgb(src),
+            Self::Aces2065_1 => Aces2065_1::to_linear_srgb(src),
             Self::AcesCg => AcesCg::to_linear_srgb(src),
             Self::XyzD50 => XyzD50::to_linear_srgb(src),
             Self::XyzD65 => XyzD65::to_linear_srgb(src),
@@ -244,6 +252,7 @@ impl ColorSpaceTag {
             Self::A98Rgb => A98Rgb::clip(src),
             Self::ProphotoRgb => ProphotoRgb::clip(src),
             Self::Rec2020 => Rec2020::clip(src),
+            Self::Aces2065_1 => Aces2065_1::clip(src),
             Self::AcesCg => AcesCg::clip(src),
             Self::XyzD50 => XyzD50::clip(src),
             Self::XyzD65 => XyzD65::clip(src),